Extract relay type into new struct
