+++ relay_test.go
@@ -66,9 +66,10 @@ func Test_RecepientsCheck(t *testing.T) {
 		},
 	}
 
+	r := &relay{}
 	for _, tt := range tc {
 		t.Run(tt.name, func(t *testing.T) {
-			checker := recipientChecker(tt.allowed, tt.denied)
+			checker := r.recipientChecker(tt.allowed, tt.denied)
 
 			for _, e := range tt.emails {
 				if err := checker(smtpd.Peer{}, e); err != nil {
